Make / redirect to /Volunteers
I realized that I shouldn't have followed the comment's advice to rename index()
to upcomingBirthdays() for, even though that's what it does, CakePHP has strong
conventions about what's what.
In particular, clicking on the Volunteers logo/link in the header goes to /Volunteers
which used to mean /Volunteers/index until I broke it by that rename.
So, instead, I've renamed it back, and added a short *new* handler that just
aliases / -> /Volunteers