Cleaner hyphen-insensitivity.
It turns out I was just misusing ->func(); it can handle arbitrary
function names, it's just that the arguments to them need to be wrapped up
in an array for some reason.
Using this I was able to move the replace() call entirely into SQL, and only
use two of them: one on the data and one on the query.