Show Emergency Contact regardless of content.
The emergency contact view was hiding itself if it had no data,
or at least, only showing itself if it had some key data,
but this was ignorant of the `emergunable` flag, and so
that flag would be ignored, and not display its corresponding
message, unless there was also some dummy data entered.
This whole thing is silly. See #18.

This patch just removes the check, since it's just tiny bit of
UI sugar anyway, and caused breakage.

Fixes #23.