Limit date range to sensible defaults.
I'm assuming we won't have volunteers under 7 years old here.
There's a UI glitch: the datepicker in Firefox isn't smart enough
to automatically jump to the valid date range.
Chrome is smart enough though.
There isn't any way to control this, besides filling in
the value ahead of time, but that would be bad for other reasons,
so I'm just going to let this slide.

Fixes #28